+++ types/protocol-mapping.d.ts
@@ -1063,13 +1063,6 @@ export namespace ProtocolMapping {
             paramsType: [Protocol.Profiler.StartPreciseCoverageRequest?];
             returnType: Protocol.Profiler.StartPreciseCoverageResponse;
         };
-        /**
-         * Enable type profile.
-         */
-        'Profiler.startTypeProfile': {
-            paramsType: [];
-            returnType: void;
-        };
         'Profiler.stop': {
             paramsType: [];
             returnType: Protocol.Profiler.StopResponse;
@@ -1083,26 +1076,12 @@ export namespace ProtocolMapping {
             returnType: void;
         };
         /**
-         * Disable type profile. Disabling releases type profile data collected so far.
-         */
-        'Profiler.stopTypeProfile': {
-            paramsType: [];
-            returnType: void;
-        };
-        /**
          * Collect coverage data for the current isolate, and resets execution counters. Precise code
          * coverage needs to have started.
          */
         'Profiler.takePreciseCoverage': {
             paramsType: [];
             returnType: Protocol.Profiler.TakePreciseCoverageResponse;
-        };
-        /**
-         * Collect type profile.
-         */
-        'Profiler.takeTypeProfile': {
-            paramsType: [];
-            returnType: Protocol.Profiler.TakeTypeProfileResponse;
         };
         /**
          * Add handler to promise with given promise object id.